+++ src/endpoints/events.rs
@@ -1,6 +1,6 @@
 // #[cfg(test)]
 // mod tests;
-mod event_file;
+mod event_api;
 
 use rocket::serde::json::Json;
 use rocket_db_pools::{
@@ -12,7 +12,7 @@ use crate::endpoints::errors::{ApiResult, ApiErrors};
 use crate::db::{self, SPS};
 
 #[get("/events/<account_id>")]
-pub async fn fetch_events(account_id: i32, mut db_conn: Connection<SPS>) -> ApiResult<Json<Vec<event_file::EventFile>>>{
+pub async fn fetch_events(account_id: i32, mut db_conn: Connection<SPS>) -> ApiResult<Json<Vec<event_api::EventFile>>>{
 
     // Checking the user account actually exists
     match sqlx::query!(
@@ -32,7 +32,7 @@ pub async fn fetch_events(account_id: i32, mut db_conn: Connection<SPS>) -> ApiR
         Err(_) => return Err(ApiErrors::NotFound("No events where found".to_string()))
     };
 
-    let events: Vec<event_file::EventFile> = db_events.iter().map(|event| event.into()).collect();
+    let events: Vec<event_api::EventFile> = db_events.iter().map(|event| event.into()).collect();
 
     Ok(Json(events))
 }
